+++ NEWS.md
@@ -4,6 +4,7 @@
 - Allow to do semi-supervised learning easily with embed_tagspace
 - Attributes attached to a model are now also restored when loading a model with starspace_load_model of type 'ruimtehol'
 - Add range.textspace to get the range of embedding similarities
+- starspace now also returns the loss before training instead of only starting from epoch 1
 
 ## CHANGES IN ruimtehol VERSION 0.1.2